+++ src/app/features/notices/notices.module.ts
@@ -1,5 +1,8 @@
 import { NgModule } from '@angular/core';
 
+import { ShareButtonsModule } from 'ngx-sharebuttons/buttons';
+import { ShareIconsModule } from 'ngx-sharebuttons/icons';
+
 import { SharedModule } from '@shared/shared.module';
 
 import { NoticesRoutingModule } from '@features/notices/notices-routing.module';
@@ -7,6 +10,8 @@ import { NoticesComponent } from '@features/notices/notices.component';
 import { NoticeViewComponent } from '@features/notices/notice-view/notice-view.component';
 import { NoticeEditComponent } from '@features/notices/notice-edit/notice-edit.component';
 import { NoticesDashboardComponent } from '@features/notices/notices-dashboard/notices-dashboard.component';
+import { NoticeItemDetailComponent } from '@features/notices/notice-item-detail/notice-item-detail.component';
+import { NoticeExpansionPanelComponent } from '@features/notices/notice-expansion-panel/notice-expansion-panel.component';
 
 @NgModule({
   declarations: [
@@ -14,13 +19,18 @@ import { NoticesDashboardComponent } from '@features/notices/notices-dashboard/n
     NoticeViewComponent,
     NoticeEditComponent,
     NoticesDashboardComponent,
+    NoticeExpansionPanelComponent,
+    NoticeItemDetailComponent,
   ],
   imports: [
     SharedModule,
-    NoticesRoutingModule
+    NoticesRoutingModule,
+    ShareButtonsModule,
+    ShareIconsModule
   ],
   exports: [
     NoticesDashboardComponent,
+    NoticeExpansionPanelComponent
   ]
 })
 export class NoticesModule { }